pyproject.toml
[project]
name = "humandesign-api"
version = "1.3.0"
description = "Human Design API"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipython==8.12.3",
    "numpy==2.1.2",
    "pandas==2.2.3",
    "python-dateutil==2.9.0.post0",
    "pytz==2024.2",
    "pyswisseph==2.10.3.2",
    "tqdm==4.66.5",
    "fastapi==0.110.0",
    "uvicorn==0.27.1",
    "pydantic==2.6.3",
    "geopy==2.4.1",
    "timezonefinder",
    "python-dotenv",
    "matplotlib",
    "svgpath2mpl",
    "Pillow"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
